lets not mislead people by making the api take double when in reality gltf timestamps are always single precision floats

+++ Runtime/Scripts/GLTFSceneExporter.cs
@@ -118,7 +118,7 @@ namespace UnityGLTF
 			string propertyName,
 			GLTFAnimation animation,
 			InterpolationType interpolationType,
-			double[] times,
+			float[] times,
 			object[] values
 		);
 	}
